Update flake to include all test deps
The build of the haskell project within the flake won't run the tests
but the check phase of the flake does run them (at least the unit and
fuzz tests).

This is due to the flake not permitting network calls in checks, so
system tests have to be skipped.

The default dev shell has all the minerals to run the system tests
though